infra/requirements.txt
aws-cdk-lib==2.86.0
constructs==10.0.0
boto3==1.40.12
botocore==1.40.12
httpx==0.27.0
pytest==8.2.2
python-dotenv==1.1.1
